Extract get-context-options.js and parser-cli-arguments.js
